Move Exposure Dashboard to EH module
