+++ src/router/index.js
@@ -7,7 +7,7 @@ import PatientManagement from '../components/PatientManagement.vue'
 import Settings from '../components/Settings.vue'
 import SystemStatus from '../views/SystemStatus.vue'
 import PatientView from '../views/PatientView.vue'
-import ExposureDashboard from '../views/ExposureDashboard.vue'
+import ExposureDashboard from '../modules/eh/components/dashboard/ExposureDashboard.vue'
 
 const routes = [
   {
@@ -46,7 +46,7 @@ const routes = [
     meta: { requiresAuth: true }
   },
   {
-    path: '/exposure',
+    path: '/eh/exposure-dashboard',
     name: 'ExposureDashboard',
     component: ExposureDashboard,
     meta: { requiresAuth: true }
